requirements.txt
alembic==1.8.1
asgiref==3.3.4
astropy==5.1
certifi==2022.9.24
charset-normalizer==2.0.12
Deprecated==1.2.13
Django==3.2.2
django-cors-headers==3.5.0
django-environ==0.4.5
django-filter==2.4.0
django-guardian==2.3.0
django-versatileimagefield==2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
et-xmlfile==1.1.0
greenlet==1.1.3
gunicorn==20.1.0
idna==3.4
install==1.3.4
Mako==1.2.3
Markdown==3.3.4
MarkupSafe==2.1.1
newrelic==5.20.1.150
numpy==1.23.3
openpyxl==3.0.7
packaging==21.3
pandas==1.2.4
pangres==2.3.1
Pillow==8.2.0
psycopg2-binary==2.8.6
pyerfa==2.0.0.1
Pygments==2.9.0
PyJWT==2.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2021.1
PyYAML==6.0
redis==4.1.1
requests==2.26.0
sentry-sdk==1.9.0
sgp4==2.21
six==1.16.0
SQLAlchemy==1.4.25
sqlparse==0.4.1
Unidecode==1.3.2
urllib3==1.26.12
websocket-client==1.4.1
whitenoise==5.2.0
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==1.4.3
